CHANGED: Tests for ClamAV scan, and Memtest scan

+++ linux/hw_tests.sh
@@ -93,28 +93,39 @@ init_memtest () {
 #	if sudo test -d "/boot/efi/EFI/memtest/oldreports"; then
 #		sudo mkdir /boot/efi/EFI/memtest/oldreports
 #	fi
-	sudo bash -c 'mv /boot/efi/EFI/memtest/*.html /boot/efi/EFI/memtest/oldreports'
 
-	$DIAG
+	if [ "$FAHT_MEMTEST" == "ON" ]; then
 
-	#FAHT_CURR_
-	save_vars
-	
-	reboot_tomemtest
+		sudo bash -c 'mv /boot/efi/EFI/memtest/*.html /boot/efi/EFI/memtest/oldreports'
+
+		$DIAG
+
+		#FAHT_CURR_
+		save_vars
+		
+		reboot_tomemtest
+	else
+		echo "Memtest not set! Skipping..."
+	fi
 }
 
 finish_memtest () {
-	echo "-------"
-	echo "Importing Memtest results"
 
-	if [ -d /sys/firmware/efi ]; then
-		export FAHT_MEMTEST_REPORT_FILE="${FAHT_WORKINGDIR}"/memtest.html
-		echo Copying HTML report file to $FAHT_WORKINGDIR
-		export MT_FILE=$(sudo bash -c 'ls /boot/efi/EFI/memtest/*.html')
-		sudo bash -c 'cp ${0} ${1}' "$MT_FILE" "$FAHT_MEMTEST_REPORT_FILE"
-		memtest_parsing
-		sudo rm $MT_FILE
-		echo Importing results
+	if [ "$FAHT_MEMTEST" == "ON" ]; then
+		echo "-------"
+		echo "Importing Memtest results"
+
+		if [ -d /sys/firmware/efi ]; then
+			export FAHT_MEMTEST_REPORT_FILE="${FAHT_WORKINGDIR}"/memtest.html
+			echo Copying HTML report file to $FAHT_WORKINGDIR
+			export MT_FILE=$(sudo bash -c 'ls /boot/efi/EFI/memtest/*.html')
+			sudo bash -c 'cp ${0} ${1}' "$MT_FILE" "$FAHT_MEMTEST_REPORT_FILE"
+			memtest_parsing
+			sudo rm $MT_FILE
+			echo Importing results
+		fi
+	else
+		echo "Memtest not set! Skipping..."
 	fi
 }
